+++ src/user_interface/mod.rs
@@ -19,13 +19,15 @@ use crate::graphics::GraphicsWindow;
 use crate::scene::SceneAction;
 use crate::user_interface::control_state::{ControlState, VisibleControl};
 use crate::user_interface::gravity::GravityMessage;
-use crate::user_interface::keyboard::{KeyboardMessage, Menu};
+use crate::user_interface::keyboard::KeyboardMessage;
+use crate::user_interface::menu::Menu;
 use crate::user_interface::strain_threshold::StrainThresholdMessage;
 
 mod strain_threshold;
 mod gravity;
 mod keyboard;
 mod control_state;
+mod menu;
 
 const FRAME_RATE_MEASURE_INTERVAL_SECS: f64 = 0.5;
 
@@ -53,74 +55,6 @@ pub enum Action {
     StartTinkering,
     ToggleDebug,
     AddBrick,
-}
-
-fn fabric_menu(fabrics: &[FabricPlan], below: Vec<String>) -> Vec<Menu> {
-    let sub_fabrics: Vec<_> = fabrics
-        .iter()
-        .filter(|fabric| {
-            let mut compare = below.clone();
-            compare.push(fabric.name.last().unwrap().clone());
-            compare == fabric.name
-        })
-        .collect();
-    if sub_fabrics.is_empty() {
-        let mut unique: Vec<String> = Vec::new();
-        for plan in fabrics {
-            let next_name = plan.name.get(below.len()).unwrap();
-            match unique.last() {
-                None => unique.push(next_name.clone()),
-                Some(last_next_name) if next_name != last_next_name => unique.push(next_name.clone()),
-                _ => {}
-            }
-        }
-        unique
-            .iter()
-            .map(|first| {
-                let mut new_below = below.clone();
-                new_below.push(first.clone());
-                Menu::new(first.as_str(), fabric_menu(fabrics, new_below))
-            })
-            .collect()
-    } else {
-        sub_fabrics
-            .into_iter()
-            .map(|fabric_plan| {
-                let label = fabric_plan.name.last().unwrap();
-                Menu::action(label.as_str(), Action::Crucible(CrucibleAction::BuildFabric(fabric_plan.clone())))
-            })
-            .collect()
-    }
-}
-
-fn speed_menu() -> Vec<Menu> {
-    [(0usize, "Paused"), (5, "Glacial"), (25, "Slow"), (125, "Normal"), (625, "Fast")]
-        .into_iter()
-        .map(|(speed, label)|
-            Menu::action(label, Action::Crucible(CrucibleAction::SetSpeed(speed))))
-        .collect()
-}
-
-fn action_menu() -> Menu {
-    Menu::new("Tensegrity Lab", vec![
-        Menu::new("Fabric", fabric_menu(&Library::standard().fabrics, Vec::new())),
-        Menu::new("Speed", speed_menu()),
-        Menu::new("Camera", vec![
-            Menu::action("Next face", Action::SelectNextFace),
-            Menu::action("Watch Fabric Midpoint", Action::Scene(SceneAction::WatchMidpoint)),
-            Menu::action("Watch Origin", Action::Scene(SceneAction::WatchOrigin)),
-        ]),
-        Menu::new("Widget", vec![
-            Menu::action("Gravity", Action::ShowControl(VisibleControl::Gravity)),
-            Menu::action("Strain threshold", Action::ShowControl(VisibleControl::StrainThreshold)),
-            Menu::action("Clear", Action::ShowControl(VisibleControl::Nothing)),
-        ]),
-        Menu::new("Etc", vec![
-            Menu::action("Tinker", Action::StartTinkering),
-            Menu::action("Debug toggle", Action::ToggleDebug),
-            Menu::action("Add brick", Action::AddBrick),
-        ]),
-    ])
 }
 
 /// Largely adapted from https://github.com/iced-rs/iced/blob/master/examples/integration_wgpu/src/main.rs
